Move InferenceAPI into new file
